Experimental changes being merged to main

+++ src/Firebase/Styles Functions/deleteStyleFromFirebase.js
@@ -1,32 +1,56 @@
+"use server";
 import { db, storage } from "../firebase";
-import { getDoc,doc, deleteDoc } from "firebase/firestore";
+import { getDoc, doc, deleteDoc } from "firebase/firestore";
 import { ref, deleteObject } from "firebase/storage";
+import { revalidatePath } from "next/cache";
 
 const deleteStyleFromDB = async (styleId) => {
-	try {
-		const docRef = doc(db, "Style", styleId);
-		const docSnapshot = await getDoc(docRef);
+  try {
+    const docRef = doc(db, "Style", styleId);
+    const docSnapshot = await getDoc(docRef);
 
-		if (docSnapshot.exists()) {
-			const data = docSnapshot.data();
-
-			if (data) {
-				const imageRef = ref(storage, `Style/${styleId}`);
-				await deleteObject(imageRef);
-			}
-
-			await deleteDoc(docRef);
-
-			return Promise.resolve(
-				"Style data and associated image deleted successfully"
-			);
-		} else {
-			return Promise.reject(`Style with ID ${styleId} does not exist`);
-		}
-	} catch (error) {
-		console.error("Firebase Error: " + error.message);
-		return Promise.reject("Firebase Error: " + error.message);
-	}
+    if (docSnapshot.exists()) {
+      let usage = docSnapshot.data().usage;
+      let usageCases = "";
+      for (const key in usage) {
+        if (usage[key] !== 0) {
+          usageCases += key.toUpperCase() + ", ";
+        }
+      }
+      if (usageCases !== "") {
+        return {
+          type: "error",
+          message: `Style cannot be deleted. It is being used in ${usageCases.slice(
+            0,
+            -2
+          )}.`,
+        };
+      } else {
+        const data = docSnapshot.data();
+        if (data) {
+          const imageRef = ref(storage, `Style/${styleId}`);
+          await deleteObject(imageRef);
+        }
+        await deleteDoc(docRef);
+        revalidatePath("/admin/roles-analytics-cities", "page");
+        return {
+          type: "success",
+          message: "Style deleted successfully.",
+        };
+      }
+    } else {
+      return {
+        type: "error",
+        message: "Something went wrong, please try again later.",
+      };
+    }
+  } catch (error) {
+    console.error("Error deleting the style: ", error);
+    return {
+      type: "error",
+      message: "Something went wrong, please try again later.",
+    };
+  }
 };
 
 export default deleteStyleFromDB;
